+++ src/classes/Main.java
@@ -22,15 +22,24 @@ import java.util.concurrent.Semaphore;
 public class Main {
 
     public static void main(String[] args) {
-        Semaphore driveGuiones = new Semaphore(11);
 
-        // Pasamos los parametros necesarios para el desarrollador de narrativa.
-        DesarrolladorNarrativa empleado = new DesarrolladorNarrativa(0, 0, /**capacidadDrive,*/ 0, 10, driveGuiones, 0, true);
-        empleado.start();
-        
-        DesarrolladorNarrativa empleado2 = new DesarrolladorNarrativa(0, 0, /**capacidadDrive,*/ 0, 10, driveGuiones, 0, true);
-        empleado2.start();
-        // FUNCIONAAA SUUU
-//        }
+        int capacidadDrive = 25;
+        Semaphore driveN = new Semaphore(capacidadDrive);
+
+        int numeroDeEmpleados = 5; 
+
+        for (int i = 1; i <= numeroDeEmpleados; i++) {
+            String nombreHilo = "empleadoN" + i;
+            DesarrolladorNarrativa empleadoN = new DesarrolladorNarrativa(0, 0, 0, 10, driveN, 0, true, 6);
+            empleadoN.setName(nombreHilo);
+            empleadoN.start();
+        }
+
     }
 }
+
+        
+        // FUNCIONAAA SUUU
+//        }
+    
+