+++ src/ChangeLog/ChangeLog.md
@@ -1,4 +1,12 @@
 # Registro de Cambios
+## [1.0.7] - 06/10/2023 (Juandb1507)
+
+- Cambios interfaz Bethesda y TabbedPane del Dashboard principal.
+- Lógica de creación de hilos.
+
+## [1.0.6] - 05/10/2023 (Sebasr22)
+
+- Se agregaron los pagos a los desarrolladores de narrativa.
 
 ## [1.0.7] - 05/10/2023 (Sebasr22)